pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["nikola.gilev <nikola.gilev@codechem.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.22"
fastapi = "^0.115.8"
sqlalchemy = "^2.0.38"
psycopg2 = "^2.9.10"
uvicorn = "^0.34.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
